requirements.txt
autopep8==1.4.4
bingmaps==0.3.7
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
geographiclib==1.50
geopy==1.20.0
googlemaps==3.1.3
marshmallow==2.6.0
mixin==1.1
Pillow==6.2.0
pycodestyle==2.5.0
pytz==2019.2
requests==2.9.1
sqlparse==0.3.0
urllib3==1.25.6
xmltodict==0.10.1
